+++ src/gui/options/option-elements-inject.ts
@@ -174,7 +174,7 @@ modmanager.gui.OptionsOptionButton = ig.GuiElementBase.extend({
 
         this.button = new sc.ButtonGui(option.name)
         if (option.onPress) {
-            this.button.onButtonPress = option.onPress
+            this.button.onButtonPress = option.onPress.bind(option)
         }
 
         // this.button.setPos(5, 8)